test: re-organise the LSP6 test suites
