+++ tests/LSP6KeyManager/Admin/PermissionChangeAddExtensions.test.ts
