+++ tests/LSP6KeyManager/Interactions/AllowedFunctions.test.ts
