+++ tests/LSP6KeyManager/Reentrancy/SingleExecuteRelayCallToSingleExecuteRelayCall.test.ts
@@ -3,12 +3,16 @@ import { ethers } from "hardhat";
 
 //types
 import { BigNumber, BytesLike } from "ethers";
+import {
+  SingleReentrancyRelayer__factory,
+  UniversalProfile__factory,
+} from "../../../types";
 
 // constants
-import { ERC725YDataKeys } from "../../../../constants";
+import { ERC725YDataKeys } from "../../../constants";
 
 // setup
-import { LSP6TestContext } from "../../../utils/context";
+import { LSP6TestContext } from "../../utils/context";
 
 // helpers
 import {
@@ -23,11 +27,11 @@ import {
   changeUniversalReceiverDelegateTestCases,
   // Functions
   generateRelayCall,
-  generateExecutePayload,
+  generateSingleRelayPayload,
   loadTestCase,
 } from "./reentrancyHelpers";
 
-export const testSingleExecuteRelayCallToSingleExecute = (
+export const testSingleExecuteRelayCallToSingleExecuteRelayCall = (
   buildContext: (initialFunding?: BigNumber) => Promise<LSP6TestContext>,
   buildReentrancyContext: (
     context: LSP6TestContext
@@ -35,10 +39,27 @@ export const testSingleExecuteRelayCallToSingleExecute = (
 ) => {
   let context: LSP6TestContext;
   let reentrancyContext: ReentrancyContext;
+  let executePayload: BytesLike;
 
   before(async () => {
     context = await buildContext(ethers.utils.parseEther("10"));
     reentrancyContext = await buildReentrancyContext(context);
+
+    const reentrantCallPayload =
+      new SingleReentrancyRelayer__factory().interface.encodeFunctionData(
+        "relayCallThatReenters",
+        [context.keyManager.address]
+      );
+    executePayload =
+      new UniversalProfile__factory().interface.encodeFunctionData(
+        "execute(uint256,address,uint256,bytes)",
+        [
+          0,
+          reentrancyContext.singleReentarncyRelayer.address,
+          0,
+          reentrantCallPayload,
+        ]
+      );
   });
 
   describe("when reentering and transferring value", () => {
@@ -48,21 +69,25 @@ export const testSingleExecuteRelayCallToSingleExecute = (
       payload: BytesLike;
     };
     before(async () => {
-      const executePayload = generateExecutePayload(
-        context.keyManager.address,
-        reentrancyContext.reentrantContract.address,
-        "TRANSFERVALUE"
-      );
-
       relayCallParams = await generateRelayCall(
         context.keyManager,
         executePayload,
         reentrancyContext.signer
       );
+
+      await generateSingleRelayPayload(
+        context.universalProfile,
+        context.keyManager,
+        "TRANSFERVALUE",
+        reentrancyContext.singleReentarncyRelayer,
+        reentrancyContext.reentrantSigner,
+        reentrancyContext.newControllerAddress,
+        reentrancyContext.newURDAddress
+      );
     });
 
     transferValueTestCases.NotAuthorised.forEach((testCase) => {
-      it(`should revert if the reentrant contract has the following permission set: PRESENT - ${
+      it(`should revert if the reentrant signer has the following permission set: PRESENT - ${
         testCase.permissionsText
       }; MISSING - ${testCase.missingPermission}; AllowedCalls - ${
         testCase.allowedCalls ? "YES" : "NO"
@@ -71,8 +96,8 @@ export const testSingleExecuteRelayCallToSingleExecute = (
           "TRANSFERVALUE",
           testCase,
           context,
-          reentrancyContext.reentrantContract.address,
-          reentrancyContext.reentrantContract.address
+          reentrancyContext.reentrantSigner.address,
+          reentrancyContext.singleReentarncyRelayer.address
         );
 
         await expect(
@@ -86,19 +111,19 @@ export const testSingleExecuteRelayCallToSingleExecute = (
         )
           .to.be.revertedWithCustomError(context.keyManager, "NotAuthorised")
           .withArgs(
-            reentrancyContext.reentrantContract.address,
+            reentrancyContext.reentrantSigner.address,
             testCase.missingPermission
           );
       });
     });
 
-    it("should revert if the reentrant contract has the following permissions: REENTRANCY, TRANSFERVALUE & NO AllowedCalls", async () => {
+    it("should revert if the reentrant signer has the following permissions: REENTRANCY, TRANSFERVALUE & NO AllowedCalls", async () => {
       await loadTestCase(
         "TRANSFERVALUE",
         transferValueTestCases.NoCallsAllowed,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.singleReentarncyRelayer.address
       );
 
       await expect(
@@ -112,13 +137,13 @@ export const testSingleExecuteRelayCallToSingleExecute = (
       ).to.be.revertedWithCustomError(context.keyManager, "NoCallsAllowed");
     });
 
-    it("should pass if the reentrant contract has the following permissions: REENTRANCY, TRANSFERVALUE & AllowedCalls", async () => {
+    it("should pass if the reentrant signer has the following permissions: REENTRANCY, TRANSFERVALUE & AllowedCalls", async () => {
       await loadTestCase(
         "TRANSFERVALUE",
         transferValueTestCases.ValidCase,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.singleReentarncyRelayer.address
       );
 
       expect(
@@ -143,7 +168,7 @@ export const testSingleExecuteRelayCallToSingleExecute = (
 
       expect(
         await context.universalProfile.provider.getBalance(
-          reentrancyContext.reentrantContract.address
+          reentrancyContext.singleReentarncyRelayer.address
         )
       ).to.equal(ethers.utils.parseEther("1"));
     });
@@ -156,21 +181,25 @@ export const testSingleExecuteRelayCallToSingleExecute = (
       payload: BytesLike;
     };
     before(async () => {
-      const executePayload = generateExecutePayload(
-        context.keyManager.address,
-        reentrancyContext.reentrantContract.address,
-        "SETDATA"
-      );
-
       relayCallParams = await generateRelayCall(
         context.keyManager,
         executePayload,
         reentrancyContext.signer
       );
+
+      await generateSingleRelayPayload(
+        context.universalProfile,
+        context.keyManager,
+        "SETDATA",
+        reentrancyContext.singleReentarncyRelayer,
+        reentrancyContext.reentrantSigner,
+        reentrancyContext.newControllerAddress,
+        reentrancyContext.newURDAddress
+      );
     });
 
     setDataTestCases.NotAuthorised.forEach((testCase) => {
-      it(`should revert if the reentrant contract has the following permission set: PRESENT - ${
+      it(`should revert if the reentrant signer has the following permission set: PRESENT - ${
         testCase.permissionsText
       }; MISSING - ${testCase.missingPermission}; AllowedERC725YDataKeys - ${
         testCase.allowedERC725YDataKeys ? "YES" : "NO"
@@ -179,8 +208,8 @@ export const testSingleExecuteRelayCallToSingleExecute = (
           "SETDATA",
           testCase,
           context,
-          reentrancyContext.reentrantContract.address,
-          reentrancyContext.reentrantContract.address
+          reentrancyContext.reentrantSigner.address,
+          reentrancyContext.singleReentarncyRelayer.address
         );
 
         await expect(
@@ -194,19 +223,19 @@ export const testSingleExecuteRelayCallToSingleExecute = (
         )
           .to.be.revertedWithCustomError(context.keyManager, "NotAuthorised")
           .withArgs(
-            reentrancyContext.reentrantContract.address,
+            reentrancyContext.reentrantSigner.address,
             testCase.missingPermission
           );
       });
     });
 
-    it("should revert if the reentrant contract has the following permissions: REENTRANCY, SETDATA & NO AllowedERC725YDataKeys", async () => {
+    it("should revert if the reentrant signer has the following permissions: REENTRANCY, SETDATA & NO AllowedERC725YDataKeys", async () => {
       await loadTestCase(
         "SETDATA",
         setDataTestCases.NoERC725YDataKeysAllowed,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.singleReentarncyRelayer.address
       );
 
       await expect(
@@ -223,13 +252,13 @@ export const testSingleExecuteRelayCallToSingleExecute = (
       );
     });
 
-    it("should pass if the reentrant contract has the following permissions: REENTRANCY, SETDATA & AllowedERC725YDataKeys", async () => {
+    it("should pass if the reentrant signer has the following permissions: REENTRANCY, SETDATA & AllowedERC725YDataKeys", async () => {
       await loadTestCase(
         "SETDATA",
         setDataTestCases.ValidCase,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.singleReentarncyRelayer.address
       );
 
       await context.keyManager
@@ -260,27 +289,31 @@ export const testSingleExecuteRelayCallToSingleExecute = (
       payload: BytesLike;
     };
     before(async () => {
-      const executePayload = generateExecutePayload(
-        context.keyManager.address,
-        reentrancyContext.reentrantContract.address,
-        "ADDCONTROLLER"
-      );
-
       relayCallParams = await generateRelayCall(
         context.keyManager,
         executePayload,
         reentrancyContext.signer
       );
+
+      await generateSingleRelayPayload(
+        context.universalProfile,
+        context.keyManager,
+        "ADDCONTROLLER",
+        reentrancyContext.singleReentarncyRelayer,
+        reentrancyContext.reentrantSigner,
+        reentrancyContext.newControllerAddress,
+        reentrancyContext.newURDAddress
+      );
     });
 
     addPermissionsTestCases.NotAuthorised.forEach((testCase) => {
-      it(`should revert if the reentrant contract has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
+      it(`should revert if the reentrant signer has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
         await loadTestCase(
           "ADDCONTROLLER",
           testCase,
           context,
-          reentrancyContext.reentrantContract.address,
-          reentrancyContext.reentrantContract.address
+          reentrancyContext.reentrantSigner.address,
+          reentrancyContext.singleReentarncyRelayer.address
         );
 
         await expect(
@@ -294,19 +327,19 @@ export const testSingleExecuteRelayCallToSingleExecute = (
         )
           .to.be.revertedWithCustomError(context.keyManager, "NotAuthorised")
           .withArgs(
-            reentrancyContext.reentrantContract.address,
+            reentrancyContext.reentrantSigner.address,
             testCase.missingPermission
           );
       });
     });
 
-    it("should pass if the reentrant contract has the following permissions: REENTRANCY, ADDCONTROLLER", async () => {
+    it("should pass if the reentrant signer has the following permissions: REENTRANCY, ADDCONTROLLER", async () => {
       await loadTestCase(
         "ADDCONTROLLER",
         addPermissionsTestCases.ValidCase,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.singleReentarncyRelayer.address
       );
 
       await context.keyManager
@@ -338,27 +371,31 @@ export const testSingleExecuteRelayCallToSingleExecute = (
       payload: BytesLike;
     };
     before(async () => {
-      const executePayload = generateExecutePayload(
-        context.keyManager.address,
-        reentrancyContext.reentrantContract.address,
-        "EDITPERMISSIONS"
-      );
-
       relayCallParams = await generateRelayCall(
         context.keyManager,
         executePayload,
         reentrancyContext.signer
       );
+
+      await generateSingleRelayPayload(
+        context.universalProfile,
+        context.keyManager,
+        "EDITPERMISSIONS",
+        reentrancyContext.singleReentarncyRelayer,
+        reentrancyContext.reentrantSigner,
+        reentrancyContext.newControllerAddress,
+        reentrancyContext.newURDAddress
+      );
     });
 
     editPermissionsTestCases.NotAuthorised.forEach((testCase) => {
-      it(`should revert if the reentrant contract has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
+      it(`should revert if the reentrant signer has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
         await loadTestCase(
           "EDITPERMISSIONS",
           testCase,
           context,
-          reentrancyContext.reentrantContract.address,
-          reentrancyContext.reentrantContract.address
+          reentrancyContext.reentrantSigner.address,
+          reentrancyContext.singleReentarncyRelayer.address
         );
 
         await expect(
@@ -372,19 +409,19 @@ export const testSingleExecuteRelayCallToSingleExecute = (
         )
           .to.be.revertedWithCustomError(context.keyManager, "NotAuthorised")
           .withArgs(
-            reentrancyContext.reentrantContract.address,
+            reentrancyContext.reentrantSigner.address,
             testCase.missingPermission
           );
       });
     });
 
-    it("should pass if the reentrant contract has the following permissions: REENTRANCY, EDITPERMISSIONS", async () => {
+    it("should pass if the reentrant signer has the following permissions: REENTRANCY, EDITPERMISSIONS", async () => {
       await loadTestCase(
         "EDITPERMISSIONS",
         editPermissionsTestCases.ValidCase,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.singleReentarncyRelayer.address
       );
 
       await context.keyManager
@@ -415,27 +452,31 @@ export const testSingleExecuteRelayCallToSingleExecute = (
       payload: BytesLike;
     };
     before(async () => {
-      const executePayload = generateExecutePayload(
-        context.keyManager.address,
-        reentrancyContext.reentrantContract.address,
-        "ADDUNIVERSALRECEIVERDELEGATE"
-      );
-
       relayCallParams = await generateRelayCall(
         context.keyManager,
         executePayload,
         reentrancyContext.signer
       );
+
+      await generateSingleRelayPayload(
+        context.universalProfile,
+        context.keyManager,
+        "ADDUNIVERSALRECEIVERDELEGATE",
+        reentrancyContext.singleReentarncyRelayer,
+        reentrancyContext.reentrantSigner,
+        reentrancyContext.newControllerAddress,
+        reentrancyContext.newURDAddress
+      );
     });
 
     addUniversalReceiverDelegateTestCases.NotAuthorised.forEach((testCase) => {
-      it(`should revert if the reentrant contract has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
+      it(`should revert if the reentrant signer has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
         await loadTestCase(
           "ADDUNIVERSALRECEIVERDELEGATE",
           testCase,
           context,
-          reentrancyContext.reentrantContract.address,
-          reentrancyContext.reentrantContract.address
+          reentrancyContext.reentrantSigner.address,
+          reentrancyContext.singleReentarncyRelayer.address
         );
 
         await expect(
@@ -449,19 +490,19 @@ export const testSingleExecuteRelayCallToSingleExecute = (
         )
           .to.be.revertedWithCustomError(context.keyManager, "NotAuthorised")
           .withArgs(
-            reentrancyContext.reentrantContract.address,
+            reentrancyContext.reentrantSigner.address,
             testCase.missingPermission
           );
       });
     });
 
-    it("should pass if the reentrant contract has the following permissions: REENTRANCY, ADDUNIVERSALRECEIVERDELEGATE", async () => {
+    it("should pass if the reentrant signer has the following permissions: REENTRANCY, ADDUNIVERSALRECEIVERDELEGATE", async () => {
       await loadTestCase(
         "ADDUNIVERSALRECEIVERDELEGATE",
         addUniversalReceiverDelegateTestCases.ValidCase,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.singleReentarncyRelayer.address
       );
 
       await context.keyManager
@@ -491,28 +532,32 @@ export const testSingleExecuteRelayCallToSingleExecute = (
       payload: BytesLike;
     };
     before(async () => {
-      const executePayload = generateExecutePayload(
-        context.keyManager.address,
-        reentrancyContext.reentrantContract.address,
-        "CHANGEUNIVERSALRECEIVERDELEGATE"
-      );
-
       relayCallParams = await generateRelayCall(
         context.keyManager,
         executePayload,
         reentrancyContext.signer
       );
+
+      await generateSingleRelayPayload(
+        context.universalProfile,
+        context.keyManager,
+        "CHANGEUNIVERSALRECEIVERDELEGATE",
+        reentrancyContext.singleReentarncyRelayer,
+        reentrancyContext.reentrantSigner,
+        reentrancyContext.newControllerAddress,
+        reentrancyContext.newURDAddress
+      );
     });
 
     changeUniversalReceiverDelegateTestCases.NotAuthorised.forEach(
       (testCase) => {
-        it(`should revert if the reentrant contract has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
+        it(`should revert if the reentrant signer has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
           await loadTestCase(
             "CHANGEUNIVERSALRECEIVERDELEGATE",
             testCase,
             context,
-            reentrancyContext.reentrantContract.address,
-            reentrancyContext.reentrantContract.address
+            reentrancyContext.reentrantSigner.address,
+            reentrancyContext.singleReentarncyRelayer.address
           );
 
           await expect(
@@ -526,20 +571,20 @@ export const testSingleExecuteRelayCallToSingleExecute = (
           )
             .to.be.revertedWithCustomError(context.keyManager, "NotAuthorised")
             .withArgs(
-              reentrancyContext.reentrantContract.address,
+              reentrancyContext.reentrantSigner.address,
               testCase.missingPermission
             );
         });
       }
     );
 
-    it("should pass if the reentrant contract has the following permissions: REENTRANCY, CHANGEUNIVERSALRECEIVERDELEGATE", async () => {
+    it("should pass if the reentrant signer has the following permissions: REENTRANCY, CHANGEUNIVERSALRECEIVERDELEGATE", async () => {
       await loadTestCase(
         "CHANGEUNIVERSALRECEIVERDELEGATE",
         changeUniversalReceiverDelegateTestCases.ValidCase,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.singleReentarncyRelayer.address
       );
 
       await context.keyManager
