+++ tests/LSP6KeyManager/Reentrancy/SingleExecuteToBatchExecuteRelayCall.test.ts
@@ -6,13 +6,13 @@ import { BigNumber, BytesLike } from "ethers";
 import {
   SingleReentrancyRelayer__factory,
   UniversalProfile__factory,
-} from "../../../../types";
+} from "../../../types";
 
 // constants
-import { ERC725YDataKeys } from "../../../../constants";
+import { ERC725YDataKeys } from "../../../constants";
 
 // setup
-import { LSP6TestContext } from "../../../utils/context";
+import { LSP6TestContext } from "../../utils/context";
 
 // helpers
 import {
