+++ tests/LSP6KeyManager/Reentrancy/reentrancyHelpers.ts
@@ -13,7 +13,7 @@ import {
   SingleReentrancyRelayer__factory,
   BatchReentrancyRelayer__factory,
   UniversalProfile__factory,
-} from "../../../../types";
+} from "../../../types";
 
 // constants
 import {
@@ -21,10 +21,10 @@ import {
   ALL_PERMISSIONS,
   PERMISSIONS,
   CALLTYPE,
-} from "../../../../constants";
+} from "../../../constants";
 
 // setup
-import { LSP6TestContext } from "../../../utils/context";
+import { LSP6TestContext } from "../../utils/context";
 
 // helpers
 import {
@@ -33,8 +33,8 @@ import {
   LOCAL_PRIVATE_KEYS,
   signLSP6ExecuteRelayCall,
   encodeCompactBytesArray,
-} from "../../../utils/helpers";
-import { setupKeyManager } from "../../../utils/fixtures";
+} from "../../utils/helpers";
+import { setupKeyManager } from "../../utils/fixtures";
 
 // Complex permission as it has AllowedCalls
 export type TransferValueTestCase = {
