+++ tests/LSP6KeyManager/SetData/PermissionSetData.test.ts
@@ -646,19 +646,19 @@ export const shouldBehaveLikePermissionSetData = (
       describe("when sending value while setting data", async () => {
         it("should revert with Key Manager error `CannotSendValueToSetData`", async () => {
           const keys = [
-            ethers.utils.keccak256(ethers.utils.toUtf8Bytes("MyFirstKey")),
-            ethers.utils.keccak256(ethers.utils.toUtf8Bytes("MySecondKey")),
-            ethers.utils.keccak256(ethers.utils.toUtf8Bytes("MyThirdKey")),
-            ethers.utils.keccak256(ethers.utils.toUtf8Bytes("MyFourthKey")),
-            ethers.utils.keccak256(ethers.utils.toUtf8Bytes("MyFifthKey")),
+            ethers.utils.keccak256(ethers.utils.toUtf8Bytes("My First Key")),
+            ethers.utils.keccak256(ethers.utils.toUtf8Bytes("My Second Key")),
+            ethers.utils.keccak256(ethers.utils.toUtf8Bytes("My Third Key")),
+            ethers.utils.keccak256(ethers.utils.toUtf8Bytes("My Fourth Key")),
+            ethers.utils.keccak256(ethers.utils.toUtf8Bytes("My Fifth Key")),
           ];
 
           const values = [
-            ethers.utils.hexlify(ethers.utils.toUtf8Bytes("aaaaaaaaaa")),
-            ethers.utils.hexlify(ethers.utils.toUtf8Bytes("bbbbbbbbbb")),
-            ethers.utils.hexlify(ethers.utils.toUtf8Bytes("cccccccccc")),
-            ethers.utils.hexlify(ethers.utils.toUtf8Bytes("dddddddddd")),
-            ethers.utils.hexlify(ethers.utils.toUtf8Bytes("eeeeeeeeee")),
+            ethers.utils.hexlify(ethers.utils.toUtf8Bytes("0xaaaaaaaaaa")),
+            ethers.utils.hexlify(ethers.utils.toUtf8Bytes("0xbbbbbbbbbb")),
+            ethers.utils.hexlify(ethers.utils.toUtf8Bytes("0xcccccccccc")),
+            ethers.utils.hexlify(ethers.utils.toUtf8Bytes("0xdddddddddd")),
+            ethers.utils.hexlify(ethers.utils.toUtf8Bytes("0xeeeeeeeeee")),
           ];
 
           let payload = context.universalProfile.interface.encodeFunctionData(
