+++ tests/LSP6KeyManager/index.ts
@@ -1,0 +1,43 @@
+// Admin
+export * from "./Admin/PermissionChangeOwner.test";
+export * from "./Admin/PermissionChangeAddExtensions.test";
+export * from "./Admin/PermissionChangeAddURD.test";
+export * from "./Admin/PermissionSign.test";
+
+// Set Permissions
+export * from "./SetPermissions/PermissionChangeAddController.test";
+export * from "./SetPermissions/SetAllowedCalls.test";
+
+// Interactions
+export * from "./Interactions/PermissionCall.test";
+export * from "./Interactions/PermissionStaticCall.test";
+export * from "./Interactions/PermissionDelegateCall.test";
+export * from "./Interactions/PermissionDeploy.test";
+export * from "./Interactions/PermissionTransferValue.test";
+export * from "./Interactions/AllowedAddresses.test";
+export * from "./Interactions/AllowedFunctions.test";
+export * from "./Interactions/AllowedStandards.test";
+
+// Relay
+export * from "./Relay/MultiChannelNonce.test";
+export * from "./Relay/ExecuteRelayCall.test";
+
+// Batch
+export * from "./Batch/BatchExecute.test";
+
+// Reentrancy
+export * from "./Reentrancy/Reentrancy.test";
+
+// Set Data
+export * from "./SetData/PermissionSetData.test";
+export * from "./SetData/AllowedERC725YDataKeys.test";
+
+// Others
+export * from "./OtherScenarios.test";
+export * from "./Security.test";
+export * from "./LSP6ControlledToken.test";
+
+// Internals (Unit Tests for internal functions)
+export * from "./internals/AllowedCalls.internal";
+export * from "./internals/AllowedERC725YDataKeys.internal";
+export * from "./internals/ReadPermissions.internal";
