--- tests/LSP6KeyManager/internals/index.ts
@@ -1,3 +1,0 @@
-export * from "./AllowedCalls.internal";
-export * from "./AllowedERC725YDataKeys.internal";
-export * from "./ReadPermissions.internal";
