--- tests/LSP6KeyManager/tests/index.ts
@@ -1,22 +1,0 @@
-export * from "./PermissionChangeOwner.test";
-export * from "./PermissionChangeAddController.test";
-export * from "./PermissionChangeAddExtensions.test";
-export * from "./PermissionChangeAddURD.test";
-export * from "./PermissionSetData.test";
-export * from "./PermissionCall.test";
-export * from "./PermissionStaticCall.test";
-export * from "./PermissionDelegateCall.test";
-export * from "./PermissionDeploy.test";
-export * from "./PermissionTransferValue.test";
-export * from "./PermissionSign.test";
-export * from "./AllowedAddresses.test";
-export * from "./AllowedFunctions.test";
-export * from "./AllowedStandards.test";
-export * from "./AllowedERC725YDataKeys.test";
-export * from "./MultiChannelNonce.test";
-export * from "./OtherScenarios.test";
-export * from "./Security.test";
-export * from "./LSP6ControlledToken.test";
-export * from "./ExecuteRelayCall.test";
-export * from "./BatchExecute.test";
-export * from "./Reentrancy.test";
